datamap: Do not truncate size_t to int in field initialization context

+++ src/game/shared/saverestore_bitstring.h
@@ -89,7 +89,7 @@ ISaveRestoreOps *GetBitstringDataOps(BITSTRING *)
 //-------------------------------------
 
 #define DEFINE_BITSTRING(name) \
-	{ FIELD_CUSTOM, #name, { offsetof(classNameTypedef,name), 0 }, 1, FTYPEDESC_SAVE, NULL, GetBitstringDataOps(&(((classNameTypedef *)0)->name)), NULL }
+	{ FIELD_CUSTOM, #name, { (int)offsetof(classNameTypedef,name), 0 }, 1, FTYPEDESC_SAVE, NULL, GetBitstringDataOps(&(((classNameTypedef *)0)->name)), NULL }
 
 #endif // SAVERESTORE_BITSTRING_H
